Moving email.js util to server-side.
