+++ src/util/index.js
@@ -7,6 +7,5 @@ export * from './cy';
 export * from './memoize';
 export * from './strings';
 export * from './cache';
-export * from './email';
 export * from './fetch';
 export * from './document';
